[ALPS08973238] thermal: fix coverity issue
[Description]
Fix Coverity for u32 interface overflow

[Test]
boot

MTK-Commit-Id: ff99061ed85be109e71560d124f6a24978e38b53

CR-Id: ALPS08973238
Feature: Thermal Management

+++ drivers/thermal/mediatek/pmic_temp.c
@@ -65,7 +65,7 @@ static int pmic_raw_to_temp(struct pmic_tz_cali_data *cali_data, int tz_id,
 	int y_curr = val;
 
 	t_current = cali_data[tz_id].intercept +
-		div_u64((cali_data[tz_id].slope1 * y_curr), (cali_data[tz_id].slope2));
+		div_u64(((uint64_t)cali_data[tz_id].slope1 * y_curr), (cali_data[tz_id].slope2));
 
 	return t_current;
 }
@@ -89,13 +89,13 @@ static void pmic_get_temp_convert_params(struct pmic_temp_info *data)
 		else
 			cali[i].slope2 = -(factor - tz_data->o_slope);
 
-		vbe_t = (-1) * div_u64((((cali[i].o_vts) * tz_data->pullup_volt)), 4096) * 1000;
+		vbe_t = (-1) * div_u64((((uint64_t)(cali[i].o_vts) * tz_data->pullup_volt)), 4096) * 1000;
 
 		if (tz_data->o_slope_sign == 0)
-			cali[i].intercept = div_u64((vbe_t * 1000), (-(factor + tz_data->o_slope * 10)));
+			cali[i].intercept = div_u64(((uint64_t)vbe_t * 1000), (-(factor + tz_data->o_slope * 10)));
 		/*0.001 degree */
 		else
-			cali[i].intercept = div_u64((vbe_t * 1000), (-(factor - tz_data->o_slope * 10)));
+			cali[i].intercept = div_u64(((uint64_t)vbe_t * 1000), (-(factor - tz_data->o_slope * 10)));
 		/*0.001 degree */
 
 		cali[i].intercept = cali[i].intercept + (tz_data->degc_cali * div_u64(1000, 2));
